Attaque groupée retirée, bannissements temporaires, administration traduite
- Mission 2 « Attaque groupée » : jamais programmée, elle supprimait la
  flotte à l'arrivée (vaisseaux perdus). Retirée des formulaires, refusée
  par le serveur ; une flotte déjà partie fait demi-tour. verband.php et
  le bouton « Associer » retirés.
- Bannissements : levée automatique à l'échéance (la durée était ignorée) ;
  page du joueur banni traduite, avec la date de fin ; autounban.php
  (logique fausse, lié nulle part) supprimé ; admin.mo chargé dans la
  langue du joueur.
- Français en dur traduit : erreur SQL, message à tous (sujet ou texte vide
  = page blanche), listes des lunes / planètes, débannissement ;
  teleport.tpl (inutilisé) supprimé.

+++ includes/functions/FlyingFleetHandler.php
@@ -40,8 +40,16 @@ function FlyingFleetHandler (&$planet) {
 				break;
 
 			case 2:
-				// Attaque groupée
-				doquery ("DELETE FROM {{table}} WHERE `fleet_id` = '". $CurrentFleet['fleet_id'] ."';", 'fleets');
+				// Attaque groupée : jamais programmée (la flotte était supprimée à l'arrivée, vaisseaux compris).
+				// Mission retirée du formulaire ; une flotte qui l'aurait reçue fait demi-tour avec son chargement.
+				if ($CurrentFleet['fleet_mess'] == 0) {
+					if ($CurrentFleet['fleet_start_time'] <= time()) {
+						doquery ("UPDATE {{table}} SET `fleet_mess` = '1' WHERE `fleet_id` = '". $CurrentFleet['fleet_id'] ."';", 'fleets');
+					}
+				} elseif ($CurrentFleet['fleet_end_time'] <= time()) {
+					RestoreFleetToPlanet ( $CurrentFleet, true );
+					doquery ("DELETE FROM {{table}} WHERE `fleet_id` = '". $CurrentFleet['fleet_id'] ."';", 'fleets');
+				}
 				break;
 
 			case 3:
